refactor: Tag 디렉토리 변경 및 적용
