+++ src/components/common/Tag/index.tsx
@@ -1,0 +1,2 @@
+export type { TagProps } from './Tag';
+export { default } from './Tag';
