+++ src/components/common/TagForm/index.test.tsx
@@ -1,0 +1,7 @@
+import Default from './TagForm';
+
+describe('components/common/TagForm/index', () => {
+  it('should defined default', () => {
+    expect(Default).toBeDefined();
+  });
+});
